+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # To-Do Task Manager for IntelliJ IDEs
 
+## [1.0.1] - 2025-03-06
+
+### Added
+
+- Saving and Loading tasks to a JSON file in /.todo
+
+### Changed
+
+- Disabled resizing the Add Task Dialog Window
+
 ## [1.0.0] - 2025-03-05
 
 ### Added
@@ -7,4 +17,5 @@
 - Basic functionality
 - Adding tasks
 
+[1.0.1]: https://github.com/markbakos/intellij-todo-now/
 [1.0.0]: https://github.com/markbakos/intellij-todo-now/